ver 0.24

+++ app/Http/Controllers/ThreadsController.php
@@ -115,6 +115,17 @@ class ThreadsController extends Controller
     public function destroy($channel, Thread $thread)
     {
 //        $thread->replies()->delete();
+
+//        if($thread->user_id != auth()->id()) {
+//            if(request()->wantsJson()) {
+//                return response(['status' => 'Permission Denied'], 403);
+//            }
+//
+//            return redirect('/login');
+//            abort(403, 'You do not have permission to do that');
+//        }
+        $this->authorize('update', $thread);
+
         $thread->delete();
 
         if(request()->wantsJson()) {
